refactor: split workflow/definitions.py god file into per-mode package
Decompose the 2747-line factory/workflow/definitions.py monolith into a
factory/workflow/definitions/ package with 16 focused modules:
- __init__.py: registry + register_all() + backward-compatible re-exports
- _shared.py: _deep_qa_subgraph(), DOC_FRESHNESS_GATE_PROMPT
- build.py, design.py, improve.py, research.py, meta.py, discover.py
- review.py, refine.py, founder.py, create.py
- skill_refine.py, doc.py, spec.py, parallel_improve.py

All existing imports (from factory.workflow.definitions import X) continue
to work unchanged. No workflow logic modified — pure structural extraction.

Closes #1077
